+++ src/Facades/FbAuth.php
@@ -7,7 +7,8 @@ use Illuminate\Support\Facades\Facade;
 /**
  * @method static string encodeEmail(string $email)
  * @method static string generateRandomCode()
- * @method static string createCode(Model $user)
+ * @method static string createCode(\Illuminate\Database\Eloquent\Model $user)
+ * @method static array{title: string, body: string} getResetPasswordNotificationKeys()
  *
  * @see \Mortezamasumi\FbAuth\FbAuth
  */