+++ src/Notifications/VerifyCodeNotification.php
@@ -28,11 +28,11 @@ class VerifyCodeNotification extends VerifyEmail
             ->line(__('fb-auth::fb-auth.verify.mail_message.line2'))
             ->line(new HtmlString('<p style="font-size: 2rem; line-height: 2.5rem; font-weight: 800; text-align: center; color: black; letter-spacing: 8px;">'.$this->code.'</p>'))
             ->line(__('fb-auth::fb-auth.verify.mail_message.timeout', [
-                'count' => (int) (config('fb-auth.otp_expiration') / 60)
+                'count' => (int) (config('fb-auth.otp_expiration') / 60),
             ]))
             ->line(__('fb-auth::fb-auth.verify.mail_message.ending'))
             ->salutation(new HtmlString(__('fb-auth::fb-auth.verify.mail_message.salutation', [
-                'name' => __(config('app.name'))
+                'name' => __(config('app.name')),
             ])));
     }
 }