+++ src/Testing/TestsFbAuth.php
@@ -2,10 +2,11 @@
 
 namespace Mortezamasumi\FbAuth\Testing;
 
+use Livewire\Component;
 use Livewire\Features\SupportTesting\Testable;
 
 /**
- * @mixin Testable
+ * @mixin Testable<Component>
  */
 class TestsFbAuth
 {