+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 All notable changes to `fb-auth` will be documented in this file.
 
-## 1.0.0 - 202X-XX-XX
+## 5.0.1 - 2026-07-09
 
--   initial release
+- Minor dependency fix
+
+## 5.0.0 - 2026-07-09
+
+- Upgrade to Filament 5
+
+## 4.4.3 - 2026-06-03
+
+- Fix validation error message on expired account
+
+## 4.4.2 - 2025-12-09
+
+- Minor fixes and test improvements
+
+## 4.2.12 - 2025-08-24
+
+- Fix OTP view in password reset flow